#77 Implementacio endpoints relacionats amb votes.

+++ vote/api.py
@@ -3,12 +3,13 @@ from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from comment.models import Comment
+from comment.models import Comment, ActionType
 from news.models import Submission
+from news.serializers import SubmissionSerializer
 from vote.models import Vote
 from vote.serializers import VoteSerializer
 
-class VotesDetailApiView(APIView):
+class VoteSubmissionApiView(APIView):
 
     def get_submission(self, submission_id):
         try:
@@ -16,18 +17,25 @@ class VotesDetailApiView(APIView):
         except Submission.DoesNotExist:
             return None
 
-    def get_comment(self,comment_id):
-        try:
-            return Comment.objects.get(id=comment_id)
-        except Comment.DoesNotExist:
-            return None
+    def get(self, request,id):
+        votes = Vote.objects.all()
+        if not votes:
+            return Response(
+                {"res": "Object with news id does not exists"},
+                status=status.HTTP_400_BAD_REQUEST
+            )
 
-    def post(self, request):
+        serializer = VoteSerializer(votes,many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    def post(self, request, id):
+        act = ActionType.objects.get(name="Submission")
+
         data = {
-            'submission': request.data.get('submission'),
-            'comment': request.data.get('comment'),
-            'type': request.data.get('type'),
-            'user': request.data.get('user'),
+            'submission': id,
+            'comment': None,
+            'type': act.id,
+            'user': request.user.id,
         }
         serializer = VoteSerializer(data=data)
         if serializer.is_valid():
@@ -37,26 +45,72 @@ class VotesDetailApiView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
-        aux = False
         submission = self.get_submission(id)
-        comment = None
         if submission is None:
-            aux = True
-            comment = self.get_comment(id)
-            if comment is None:
-                return Response(
-                    {"res": "Object with todo id does not exists"},
-                    status=status.HTTP_400_BAD_REQUEST
-                )
-        if not aux:
-            vote = Vote.objects.get(submission=submission)
-        else :
-            vote = Vote.objects.get(comment=comment)
+            return Response(
+                {"res": "Object with this id does not exists"},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        vote = Vote.objects.get(submission=submission)
+        if vote is None:
+            return Response(
+                {"res": "Vote doesn't exist"},
+                status=status.HTTP_400_BAD_REQUEST
+            )
         vote.delete()
         return Response(
             {"res": "Object deleted!"},
             status=status.HTTP_200_OK
         )
+
+class VoteCommentApiView(APIView):
+    def get_comment(self, comment_id):
+        try:
+            return Comment.objects.get(id=comment_id)
+        except Comment.DoesNotExist:
+            return None
+
+    def post(self, request, id):
+        act = ActionType.objects.get(name="Comment")
+        data = {
+            'submission': None,
+            'comment': id,
+            'type': act.id,
+            'user': request.user.id,
+        }
+        serializer = VoteSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, id):
+        comment = self.get_comment(id)
+        if comment is None:
+            return Response(
+                {"res": "Object with this id does not exists"},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        vote = Vote.objects.get(comment=comment)
+        if vote is None:
+            return Response(
+                {"res": "Vote doesn't exist"},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        vote.delete()
+        return Response(
+            {"res": "Object deleted!"},
+            status=status.HTTP_200_OK
+        )
+
+
+
+
+
+
+
+
